+++ app/src/core/navigation/index.ts
@@ -1,0 +1,3 @@
+export { default as UserMenu } from './UserMenu'
+export { useNavigationItems } from './useNavigationItems'
+export type { NavItem, NavSection, AdminGroup } from './types'